spliting out GotoDialog into its own file
