+++ Lib/defconQt/gotoDialog.py
@@ -1,0 +1,93 @@
+from PyQt5.QtCore import Qt, QEvent
+from PyQt5.QtWidgets import (
+    QDialog, QGridLayout, QLabel, QRadioButton,
+    QLineEdit, QDialogButtonBox, QListWidget)
+
+class GotoDialog(QDialog):
+    alphabetical = [
+        dict(type="alphabetical", allowPseudoUnicode=True)
+    ]
+
+    def __init__(self, currentGlyph, parent=None):
+        super(GotoDialog, self).__init__(parent)
+        self.setWindowModality(Qt.WindowModal)
+        self.setWindowTitle("Go to…")
+        self.font = currentGlyph.getParent()
+        self._sortedGlyphs = self.font.unicodeData.sortGlyphNames(
+            self.font.keys(), self.alphabetical)
+
+        layout = QGridLayout(self)
+        self.glyphLabel = QLabel("Glyph:", self)
+        self.glyphEdit = QLineEdit(self)
+        self.glyphEdit.textChanged.connect(self.updateGlyphList)
+        self.glyphEdit.event = self.lineEvent
+        self.glyphEdit.keyPressEvent = self.lineKeyPressEvent
+
+        self.beginsWithBox = QRadioButton("Begins with", self)
+        self.containsBox = QRadioButton("Contains", self)
+        self.beginsWithBox.setChecked(True)
+        self.beginsWithBox.toggled.connect(self.updateGlyphList)
+
+        self.glyphList = QListWidget(self)
+        self.glyphList.itemDoubleClicked.connect(self.accept)
+
+        buttonBox = QDialogButtonBox(
+            QDialogButtonBox.Ok | QDialogButtonBox.Cancel)
+        buttonBox.accepted.connect(self.accept)
+        buttonBox.rejected.connect(self.reject)
+
+        l = 0
+        layout.addWidget(self.glyphLabel, l, 0, 1, 2)
+        layout.addWidget(self.glyphEdit, l, 2, 1, 4)
+        l += 1
+        layout.addWidget(self.beginsWithBox, l, 0, 1, 3)
+        layout.addWidget(self.containsBox, l, 3, 1, 3)
+        l += 1
+        layout.addWidget(self.glyphList, l, 0, 1, 6)
+        l += 1
+        layout.addWidget(buttonBox, l, 0, 1, 6)
+        self.setLayout(layout)
+        self.updateGlyphList(True)
+
+    def lineEvent(self, event):
+        if event.type() == QEvent.KeyPress and event.key() == Qt.Key_Tab:
+            if self.beginsWithBox.isChecked():
+                self.containsBox.toggle()
+            else:
+                self.beginsWithBox.toggle()
+            return True
+        else:
+            return QLineEdit.event(self.glyphEdit, event)
+
+    def lineKeyPressEvent(self, event):
+        key = event.key()
+        if key == Qt.Key_Up or key == Qt.Key_Down:
+            self.glyphList.keyPressEvent(event)
+        else:
+            QLineEdit.keyPressEvent(self.glyphEdit, event)
+
+    def updateGlyphList(self, select):
+        self.glyphList.clear()
+        if not self.glyphEdit.isModified():
+            self.glyphList.addItems(self._sortedGlyphs)
+        text = self.glyphEdit.text()
+        if select:
+            glyphs = [glyph for glyph in self._sortedGlyphs
+                      if glyph.startswith(text)]
+        else:
+            glyphs = [glyph for glyph in self._sortedGlyphs if text in glyph]
+        self.glyphList.addItems(glyphs)
+        if select:
+            self.glyphList.setCurrentRow(0)
+
+    @classmethod
+    def getNewGlyph(cls, parent, currentGlyph):
+        dialog = cls(currentGlyph, parent)
+        result = dialog.exec_()
+        currentItem = dialog.glyphList.currentItem()
+        newGlyph = None
+        if currentItem is not None:
+            newGlyphName = currentItem.text()
+            if newGlyphName in dialog.font:
+                newGlyph = dialog.font[newGlyphName]
+        return (newGlyph, result)
